made signup and login hooks to make code more modular
