+++ frontend/src/hooks/useLogin.js
@@ -1,0 +1,15 @@
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { login } from '../lib/api';
+
+const useLogin = () => {
+  const queryClient = useQueryClient();
+
+  const {mutate, isPending, error} = useMutation({
+    mutationFn:login,
+    onSuccess:()=>queryClient.invalidateQueries({queryKey:["authUser"]})
+  });
+
+  return { error, isPending, loginMutation: mutate};
+};
+
+export default useLogin
